backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Data validation and serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI and NLP
openai>=1.0.0
requests>=2.28.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Date and time
python-dateutil==2.8.2

# File handling
python-magic==0.4.27

# Task queue (for background tasks)
celery==5.3.4
redis==5.0.1

# Email
fastapi-mail==1.4.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0